+++ soh/soh/OTRGlobals.cpp
@@ -1786,7 +1786,7 @@ extern "C" ShopItemIdentity Randomizer_IdentifyShopItem(s32 sceneNum, u8 slotInd
     return OTRGlobals::Instance->gRandomizer->IdentifyShopItem(sceneNum, slotIndex);
 }
 
-extern "C" CowIdentity Randomizer_IdentifyCow(s32 sceneNum, s16 posX, s16 posZ) {
+extern "C" CowIdentity Randomizer_IdentifyCow(s32 sceneNum, s32 posX, s32 posZ) {
     return OTRGlobals::Instance->gRandomizer->IdentifyCow(sceneNum, posX, posZ);
 }
